requirements.txt
msgpack==0.6.1
pycryptodome==3.9.0
pymongo==3.8.0
PyYAML==5.1.1
requests==2.22.0
